Split iteraction tests to separate path
